requirements.txt
pandas
numpy
matplotlib
ipykernel
agno
openai
python-dotenv
sqlalchemy
ddgs
pypdf
chonkie
chonkie[openai]
lancedb
fastapi
googlesearch-python
arxiv